% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperspectaculR-package.R
\docType{package}
\name{hyperspectaculR-package}
\alias{hyperspectaculR}
\alias{hyperspectaculR-package}
\title{hyperspectaculR: Publication-Grade Artistic Visualization of Hyperspectral Imagery}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Creates artistic compositions and descriptive summaries from hyperspectral imaging cubes or numeric arrays. Provides radial spectral mandalas, cumulative spectral change, multi-band fusions, exact spectral density histograms, and experimental RMS spectral slope and within-pixel spectral quartiles. Returns composable 'ggplot' objects with explicit validity rules, display domains and compact original-rendering provenance. Does not read recordings, calibrate input values or infer physiology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/hyperspectaculR}
  \item \url{https://cttir.github.io/hyperspectaculR/}
  \item Report bugs at \url{https://github.com/CTTIR/hyperspectaculR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item R. Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]
}

}
\keyword{internal}
